backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
supabase>=2.0.0
